requirements.txt
itsdangerous
flask==2.2.5
Flask-Login
Flask-SQLAlchemy
Pillow
Werkzeug==2.2.2
WTForms
pandas
mysql-connector-python
pymysql
Flask-WTF
cryptography
fastapi
uvicorn
requests
shapely
httpx<0.28
python-jose